+++ crates/core/moon/src/lib.rs
@@ -140,6 +140,7 @@ pub async fn create_project_graph_context(workspace: &Workspace) -> ProjectGraph
         inherited_tasks: &workspace.tasks_config,
         toolchain_config: &workspace.toolchain_config,
         vcs: Some(&workspace.vcs),
+        working_dir: &workspace.working_dir,
         workspace_config: &workspace.config,
         workspace_root: &workspace.root,
     };