+++ nextgen/target/src/lib.rs
@@ -1,7 +1,9 @@
 mod target;
 mod target_error;
+mod target_locator;
 mod target_scope;
 
 pub use target::*;
 pub use target_error::*;
+pub use target_locator::*;
 pub use target_scope::*;